Fine tuning css. moved stars module folder to components folder. adjusted some imports.
