+++ client/src/components/related_products/Carousel.css
@@ -50,6 +50,7 @@
   justify-content: center;
   transform: translateX(calc(var(--slider2-index) * -100%));
   transition: transform .5s;
+
 }
 
 .addToOutfitBtn {
@@ -57,7 +58,7 @@
   width: 220px;
   display: flex;
   position: absolute;
-  z-index: 6;
+  z-index: 3;
 }
 
 .carouselButton {
