+++ client/src/components/reviews/ReviewTile.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import parse from '../../parse';
-import Stars from '../related_products/stars_module/Stars.jsx';
+import Stars from '../stars_module/Stars.jsx';
 import './reviewStyle.css';
 
 function ReviewTile({ review }) {
