+++ client/src/components/stars_module/SingleStar.jsx
