+++ client/src/components/stars_module/Stars.jsx
