+++ client/src/components/stars_module/__test__/Stars.spec.js
